backend/requirements.txt
fastapi==0.115.0
apscheduler==3.10.4
pymupdf>=1.24.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
openai>=1.30.0
playwright==1.47.0
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
python-docx
reportlab
